+++ {{cookiecutter.github_repository}}/settings/common.py
@@ -5,7 +5,7 @@ see: https://docs.djangoproject.com/en/dev/ref/settings/
 """
 
 # Third Party Stuff
-{%- if cookiecutter.use_sentry_for_error_reporting == "y" %}
+{%- if cookiecutter.use_sentry_for_error_reporting == 'y' %}
 import os
 {%- endif %}
 import environ
@@ -36,11 +36,15 @@ INSTALLED_APPS = (
     '{{ cookiecutter.main_module }}.users',
 
     'rest_framework',  # http://www.django-rest-framework.org/
+    'rest_framework_swagger',
     'versatileimagefield',  # https://github.com/WGBH/django-versatileimagefield/
 {%- if cookiecutter.webpack.lower() == 'y' %}
     'webpack_loader',  # https://github.com/owais/django-webpack-loader
 {%- endif %}
-{%- if cookiecutter.use_sentry_for_error_reporting == "y" %}
+{%- if cookiecutter.add_django_cors_headers.lower() == 'y' %}
+    'corsheaders',   # https://github.com/ottoyiu/django-cors-headers/
+{%- endif %}
+{%- if cookiecutter.use_sentry_for_error_reporting == 'y' %}
     'raven.contrib.django.raven_compat',
 {%- endif %}
 )
@@ -51,7 +55,7 @@ INSTALLED_APPS = (
 # django.contrib.auth
 # ------------------------------------------------------------------------------
 AUTH_USER_MODEL = 'users.User'
-AUTHENTICATION_BACKENDS = ("django.contrib.auth.backends.ModelBackend",)
+AUTHENTICATION_BACKENDS = ('django.contrib.auth.backends.ModelBackend',)
 
 PASSWORD_HASHERS = [
     'django.contrib.auth.hashers.Argon2PasswordHasher',
@@ -59,6 +63,13 @@ PASSWORD_HASHERS = [
     'django.contrib.auth.hashers.PBKDF2SHA1PasswordHasher',
     'django.contrib.auth.hashers.BCryptSHA256PasswordHasher',
     'django.contrib.auth.hashers.BCryptPasswordHasher',
+]
+
+AUTH_PASSWORD_VALIDATORS = [
+    {'NAME': 'django.contrib.auth.password_validation.UserAttributeSimilarityValidator', },
+    {'NAME': 'django.contrib.auth.password_validation.MinimumLengthValidator', 'OPTIONS': {'min_length': 6, }},
+    {'NAME': 'django.contrib.auth.password_validation.CommonPasswordValidator', },
+    {'NAME': 'django.contrib.auth.password_validation.NumericPasswordValidator', },
 ]
 
 # For Exposing browsable api urls. By default urls won't be exposed.
@@ -100,13 +111,13 @@ REST_FRAMEWORK = {
         # Mainly used for api debug.
         'rest_framework.authentication.SessionAuthentication',
     ),
-    "EXCEPTION_HANDLER": "{{ cookiecutter.main_module }}.base.exceptions.exception_handler",
+    'EXCEPTION_HANDLER': '{{ cookiecutter.main_module }}.base.exceptions.exception_handler',
 }
 # DJANGO_SITES
 # ------------------------------------------------------------------------------
 # see: http://django-sites.readthedocs.org
 SITES = {
-    "local": {"domain": "localhost:8000", "scheme": "http", "name": "localhost"},
+    'local': {'domain': 'localhost:8000', 'scheme': 'http', 'name': 'localhost'},
 }
 SITE_ID = 'local'
 
@@ -115,7 +126,10 @@ SITE_ID = 'local'
 # List of middleware classes to use.  Order is important; in the request phase,
 # this middleware classes will be applied in the order given, and in the
 # response phase the middleware will be applied in reverse order.
-MIDDLEWARE_CLASSES = (
+MIDDLEWARE = [
+{%- if cookiecutter.add_django_cors_headers.lower() == 'y' %}
+    'corsheaders.middleware.CorsMiddleware',
+{%- endif %}
     'log_request_id.middleware.RequestIDMiddleware',  # For generating/adding Request id for all the logs
     'django.middleware.security.SecurityMiddleware',
     'whitenoise.middleware.WhiteNoiseMiddleware',
@@ -125,7 +139,7 @@ MIDDLEWARE_CLASSES = (
     'django.contrib.auth.middleware.AuthenticationMiddleware',
     'django.contrib.messages.middleware.MessageMiddleware',
     'django.middleware.clickjacking.XFrameOptionsMiddleware',
-)
+]
 
 # DJANGO CORE
 # ------------------------------------------------------------------------------
@@ -149,12 +163,12 @@ LANGUAGE_CODE = 'en-us'
 
 # Languages we provide translations for
 LANGUAGES = (
-    ("en", _("English")),
+    ('en', _('English')),
 )
 
 # A tuple of directories where Django looks for translation files.
 LOCALE_PATHS = (
-    str(APPS_DIR.path("locale")),
+    str(APPS_DIR.path('locale')),
 )
 
 # If you set this to False, Django will make some optimizations so as not
@@ -196,7 +210,7 @@ EMAIL_BACKEND = env('DJANGO_EMAIL_BACKEND',
 # ------------------------------------------------------------------------------
 # See: https://docs.djangoproject.com/en/dev/ref/settings/#databases
 DATABASES = {
-    'default': env.db("DATABASE_URL", default="postgres://localhost/{{ cookiecutter.main_module }}"),
+    'default': env.db('DATABASE_URL', default='postgres://localhost/{{ cookiecutter.main_module }}'),
 }
 DATABASES['default']['ATOMIC_REQUESTS'] = True
 DATABASES['default']['CONN_MAX_AGE'] = 10
@@ -234,17 +248,17 @@ TEMPLATES = [
     },
 ]
 
-CSRF_FAILURE_VIEW = "{{ cookiecutter.main_module }}.base.views.csrf_failure"
+CSRF_FAILURE_VIEW = '{{ cookiecutter.main_module }}.base.views.csrf_failure'
 
 # STATIC FILE CONFIGURATION
 # -----------------------------------------------------------------------------
 # Absolute path to the directory static files should be collected to.
-# Example: "/var/www/example.com/static/"
+# Example: '/var/www/example.com/static/'
 # See: https://docs.djangoproject.com/en/dev/ref/settings/#static-root
 STATIC_ROOT = str(ROOT_DIR.path('.staticfiles'))
 
 # URL that handles the static files served from STATIC_ROOT.
-# Example: "http://example.com/static/", "http://static.example.com/"
+# Example: 'http://example.com/static/', 'http://static.example.com/'
 # See: https://docs.djangoproject.com/en/dev/ref/settings/#static-url
 STATIC_URL = '/static/'
 
@@ -268,12 +282,12 @@ STATICFILES_FINDERS = (
 # ------------------------------------------------------------------------------
 
 # Absolute filesystem path to the directory that will hold user-uploaded files.
-# Example: "/var/www/example.com/media/"
+# Example: '/var/www/example.com/media/'
 # See: https://docs.djangoproject.com/en/dev/ref/settings/#media-root
 MEDIA_ROOT = str(ROOT_DIR.path('.media'))
 
 # URL that handles the media served from MEDIA_ROOT.
-# Examples: "http://example.com/media/", "http://media.example.com/"
+# Examples: 'http://example.com/media/', 'http://media.example.com/'
 # See: https://docs.djangoproject.com/en/dev/ref/settings/#media-url
 MEDIA_URL = '/media/'
 
@@ -287,7 +301,19 @@ SECURE_BROWSER_XSS_FILTER = True
 X_FRAME_OPTIONS = 'DENY'
 
 # django-log-request-id - Sending request id in response
-REQUEST_ID_RESPONSE_HEADER = "REQUEST_ID"
+REQUEST_ID_RESPONSE_HEADER = 'REQUEST_ID'
+
+{%- if cookiecutter.add_celery.lower() == 'y' %}
+
+# DJANGO CELERY CONFIGURATION
+# -----------------------------------------------------------------------------
+# see: http://celery.readthedocs.org/en/latest/userguide/tasks.html#task-states
+CELERY_BROKER_URL = env('REDIS_URL', default="redis://localhost:6379/0")
+CELERY_ACCEPT_CONTENT = ['json']
+CELERY_TASK_SERIALIZER = 'json'
+CELERY_RESULT_SERIALIZER = 'json'
+CELERY_TIMEZONE = env('CELERY_TIMEZONE', default=TIME_ZONE)  # Use django's timezone by default
+{%- endif %}
 
 # LOGGING CONFIGURATION
 # ------------------------------------------------------------------------------
@@ -371,12 +397,12 @@ LOGGING = {
 
 def get_release():
     import {{cookiecutter.main_module}}
-    {%- if cookiecutter.use_sentry_for_error_reporting == "y" %}
+    {%- if cookiecutter.use_sentry_for_error_reporting == 'y' %}
     import raven
     from raven import exceptions as raven_exceptions
     {%- endif %}
     release = {{cookiecutter.main_module}}.__version__
-    {%- if cookiecutter.use_sentry_for_error_reporting == "y" %}
+    {%- if cookiecutter.use_sentry_for_error_reporting == 'y' %}
     try:
         git_hash = raven.fetch_git_sha(os.path.dirname(os.pardir))[:7]
         release = '{}-{}'.format(release, git_hash)
@@ -389,7 +415,7 @@ def get_release():
 RELEASE_VERSION = get_release()
 
 
-{%- if cookiecutter.use_sentry_for_error_reporting == "y" %}
+{%- if cookiecutter.use_sentry_for_error_reporting == 'y' %}
 RAVEN_CONFIG = {
     'dsn': env('SENTRY_DSN', default=''),
     'environment': env('SENTRY_ENVIRONMENT', default='production'),
@@ -399,19 +425,21 @@ RAVEN_CONFIG = {
 
 SITE_INFO = {
     'RELEASE_VERSION': RELEASE_VERSION,
+{%- if cookiecutter.use_sentry_for_error_reporting == 'y' %}
     'IS_RAVEN_INSTALLED': RAVEN_CONFIG['dsn'] is not ''
+{%- endif %}
 }
 {%- if cookiecutter.webpack.lower() == 'y' %}
 
 # WEBPACK
 WEBPACK_LOADER = {
-   'DEFAULT': {
-       'CACHE': not DEBUG,
-       'BUNDLE_DIR_NAME': 'dist/assets/',  # It will add static path before and it must end with slash
-       'STATS_FILE': str(ROOT_DIR.path('webpack-stats.json')),
-       'POLL_INTERVAL': 0.1,
-       'TIMEOUT': None,
-       'IGNORE': ['.+\.hot-update.js', '.+\.map']
-   }
+    'DEFAULT': {
+        'CACHE': not DEBUG,
+        'BUNDLE_DIR_NAME': 'dist/assets/',  # It will add static path before and it must end with slash
+        'STATS_FILE': str(ROOT_DIR.path('webpack-stats.json')),
+        'POLL_INTERVAL': 0.1,
+        'TIMEOUT': None,
+        'IGNORE': ['.+\.hot-update.js', '.+\.map']
+    }
 }
 {%- endif %}