+++ {{cookiecutter.github_repository}}/wsgi.py
@@ -1,4 +1,3 @@
-# -*- coding: utf-8 -*-
 """WSGI config for  project.
 
 This module contains the WSGI application used by Django's development server
@@ -20,11 +19,19 @@ import os
 from django.core.wsgi import get_wsgi_application
 from dotenv import load_dotenv
 
-
 # Read .env file and set key/value inside it as environement variables
 # see: http://github.com/theskumar/python-dotenv
 load_dotenv(os.path.join(os.path.dirname(__file__), '.env'))
 
+# We defer to a DJANGO_SETTINGS_MODULE already in the environment. This breaks
+# if running multiple sites in the same mod_wsgi process. To fix this, use
+# mod_wsgi daemon mode with each site in its own daemon process, or use
+# os.environ['DJANGO_SETTINGS_MODULE'] = '.settings'
+os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'settings.production')
+
+# This application object is used by any WSGI server configured to use this
+# file. This includes Django's development server, if the WSGI_APPLICATION
+# setting points here.
 application = get_wsgi_application()
 
 # Apply WSGI middleware here.